remove copied code nd share it
